A lot of minor changes towards a working OSX build. The game binary compiles/links now, the gl plugin, dedicated server etc. still need to be fixed.
The Makefile has been changed to be usable for OSX builds as well (next to linux builds) and a directory unix/ was added to contain code that is shared between several *nix style builds (linux/osx for now, but possibly irix/solaris later as well).

N.B. It does not work yet! It just compiles :)