Catching up, since I was just given cvs access.
1. Nations actually care about whether you fight UFOs or not.  This is
reflected in how much they pay you.  They care the most about alien attacks
on their territory.  Nations with higher alienFriendly values are more
tolerant of failure to stop attacks and are quicker to become happy after
successes.  Half of the "happiness" reward from a victory is simply the
victory itself.  The rest is half based on what percentage of the civilians
survive, and the rest based on what percentage of the aliens were killed.

This modifies the savefiles to be one version higher, since it has to save
how happy the countries were with you when you quit.  Otherwise, you could
just save and reload, and they'd be back to normal.

2. In response to comments on the message board, I patched it so the country
Australia is renamed "Oceania" and "Arab Emirates" is renamed "United Arab
Republic".  "Oceania" is the proper name for Australia and the surrounding
islands (which would undoubtedly object to being called "Australia").  An
"Emirate" is a nation run by an emir, which would be steps backwards, be
somewhat incompatable with our future worlds, and require emirs running
nations that have no history of emirs.  On the other hand, pan-arabists have
long been trying to establish a unified arab republic (in fact, Egypt and
Syria did briefly before a coup overthrew the Syrian government; Egypt went
by the UAR until 1971). 

I will update the wiki immediately.