* Fix testall crash - cvar developer use-after-free.
This patch NULLs developer before Cvars and memory are destroyed.
An example of what happens when it's not:

Variable cvar_t *developer in src/game/g_main starts off as NULL and several
test suites never touch it.  Then developer is set non-NULL by the routing
test suite:
  testMoveEntities->SV_Map->SV_InitGame->SV_InitGameProgs->G_Init
At the end of the suite, Cvars are nuked by
  UFO_CleanSuiteRouting->TEST_Shutdown->Cvar_Shutdown
but developer still points to allocated memory until
  UFO_CleanSuiteRouting->TEST_Shutdown->Mem_Shutdown
When the Inventory suite runs, developer is checked for non-null and
dereferenced while printing a perfectly routine message at
  UFO_Init_SuiteInventory->TEST_Init->SV_Init->SV_MapcycleInit->
                           SV_ParseMapcycle->SV_MapcycleAdd->
                           Com_DBPrintf
but developer now points to an invalid non-NULL memory location and
so the program segfaults.

Signed-off-by: drane <drane@users.sourceforge.net>