* disable the intel hack by default
heisenbug:

The bug: the game is just a completely black screen, apart from the green bar during the first loading screen. The buttons in the main menu work, if you remember their position, but the screen is simply a black rectangle.

The cause: with the release of version 9.1 of the Mesa OpenGL at the beginning of the year there have been changes in order to extend the support for textures compression with the algorithm S3TC. Unfortunately none of the distributions out there support on-line compression and decompression out of the box because of patent issues in the US, so it is required to manually install an external, optional, library that implement the algorithm.

The solution: I can play again after executing one of these two actions:
a. set to 0 the entries r_intel_hack and r_ext_s3tc_compression in config.cfg
b. install the libtxc_dxtn library

Notice that libtxc_dxtn is available in Fedora and in Ubuntu, while Debian provides a (somewhat) compatible patent-free implementation called S2TC with the library libtxc-dxtn-s2tc. I'm currently running Fedora 18, and UFOAI v2.5-dev from git, but since the cause is external to the game the bug probably happens with version 2.4 too.