* Better compile error for unsupported build target platforms.
When gehmehgeh tried compiling on FreeBSD, he was getting a compile error about Com_DPrintf() being declared twice.  The real problem was worlds away, in config.h.  With this patch, the build will explain the problem correctly.  And if this #error isn't 100% correct, at least it'll be easy to debug and fix.

Signed-off-by: drane <drane@users.sourceforge.net>