Fix remembering of radar overlay state on geomap
The radar overlay remembering is done on the wrong level:
the `GEO_SetOverlay_f()` console command instead of the
underlying `GEO_SetOverlay()`. Move this handling therein.

To reproduce: in a campaign game, on geoscape, toggle the
radar overlay view on, then click on the globe, or selectr
a base and exit back to geoscape to find that the radar
overlay has been disabled back instead of being properly
remembered.